reorganizing things. new text for brew rotation
